add module fragmentation
